Player v2
This is a separate way to set up a player's inputs as well as animations. I believe it is cleaner and will be better for future updates. Only concerns is collision as well as my sprite sheet was too small so there is image clipping on the edges. Animations and character model is still just a placeholder.